requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
evaluate>=0.4.0
librosa>=0.10.0
torchcodec